Don't exceed quoted amount. Replace EmailBodyService with Razor.Templating.Core
