+++ Harvest.Core/Harvest.Core.csproj
@@ -7,6 +7,10 @@
   <ItemGroup>
     <FrameworkReference Include="Microsoft.AspNetCore.App" />
   </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Harvest.Email\Harvest.Email.csproj" />
+  </ItemGroup>
   
   <ItemGroup>
     <PackageReference Include="Azure.Storage.Blobs" Version="12.9.0" />
