+++ Harvest.Email/Harvest.Email.csproj
@@ -8,7 +8,9 @@
   <ItemGroup>
     <FrameworkReference Include="Microsoft.AspNetCore.App" />
   </ItemGroup>
-  
 
+  <ItemGroup>
+    <PackageReference Include="Razor.Templating.Core" Version="1.6.0-rc.2" />
+  </ItemGroup>
 
 </Project>
