+++ Harvest.Web/Harvest.Web.csproj
@@ -47,7 +47,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Harvest.Core\Harvest.Core.csproj" />
-    <ProjectReference Include="..\Harvest.Email\Harvest.Email.csproj" />
   </ItemGroup>
 
   <Target Name="DebugEnsureNodeEnv" BeforeTargets="Build" Condition=" '$(Configuration)' == 'Debug' And !Exists('$(SpaRoot)node_modules') ">
